requirements.txt
PyPDF2 
pinecone-client 
sentence-transformers 
cohere
streamlit
numpy
typing
dotenv
-e .